Implementation of an early stopping criteria into the training checking for no improvement (aka convergence since we do not observe overfitting). This is implemented such as none of the metrics in validation set are getting better.
Modification on the function splitting the data into training, validation and test, since if the data is preprocessed, the "final_splits.json" file is not created and therefore the training fails. If self.fold is a number that number is considered to be the fold (5 folds are created) for testing, the next fold is considered for validation (if fold number 5 is chosen for testing then 0 is chosen for validation) and the other three are considered for training. On the other hand, if "train_val_test" is specified as self.fold, that will consider there is an existing split done before for the dataset that shall be reused for reproducability.

Modification in the get_dataloader function to obtain the shape of the data in order to be inserted in the generation of the neural network.

Modified hook function obtaining the gradients only of the parameters of the network.

Modified logging of gradients to WandB to only log the linear layer parameters when training for UNet (might need modification in the future to generalize to the other models)